+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Types of changes:
 
 * Convert jsonl file to `datasets.DatasetDict` for processing with `transformers` library
 * Function to train a model on the `datasets.DatasetDict` using a pretrained model
+* Log progress when running `get_text_snippets` every 1000 records
 
 ### Changed